fix(contracts): the resolver could not read a shared-path citation (Refs #298)
#298's fix took the measured citation set from 77 to 211, but left one shape the
corpus actually uses unread. Where a contract names several items under one path:

    src/core/tests_webhook_http.rs::every_response_body_is_valid_json,
    response_escapes_hostile_codes;

`item_after` returns the FIRST item and stops at the comma. Ten function names
in the corpus are written as continuations and were resolved by nothing, so a
citation naming a function that exists nowhere passed this suite green. That is
#298's own thesis one level up: a guard reporting a total it did not measure.

`items_after` continues consuming comma-separated bare identifiers under the
same path, stopping at `;`, at any token carrying `.` or `:` (a new path or a
qualified item), and at anything that is not a bare Rust identifier.

VERIFIED BY POISONING THE CORPUS, not by asserting the parser should cope.
Replacing `response_escapes_hostile_codes` in webhook-receiver-v1.yaml with a
name present nowhere in the tree:

  before   7 passed; 0 failed                      <- the hole
  after    6 passed; 1 FAILED
      every_cited_item_is_in_the_file_that_is_cited
      ... cites `src/core/tests_webhook_http.rs::this_function_does_not_exist_anywhere`
      — it exists nowhere in src/, tests/ or benches/

Run TWICE: once before extracting the parser and once after, because a refactor
that quietly broke the parser would look exactly like success. Corpus restored
both times; 7 passed, 0 failed.

The parser moves to tests/common/contract_citations.rs because the addition took
the test file to 545 lines and the repo's ceiling is 500 — the same `#[path]`
pattern tests/common/budget_harness.rs already uses. It earns its own file: it
decides what the corpus is understood to be CLAIMING, and every count this suite
reports is downstream of it. 389 and 165 lines.

STILL OPEN, NOT FIXED HERE: `the_parser_reads_every_citation_shape_the_corpus_uses`
claims a totality it does not have. Its ten cases are hand-written literals, not
derived from the corpus, and none was a comma list — which is why it stayed green
throughout while the parser was blind to a shape the corpus uses on ten lines.
Deriving those cases from the corpus is the honest fix and is its own change.
